split question reproductions
