+++ .github/actions/build-e2e-matrix/action.yml
@@ -45,6 +45,7 @@ runs:
             ["organization", {} ],
             ["permissions", {} ],
             ["question", {} ],
+            ["question-reproductions", {} ],
             ["search", {}],
             ["sharing", {} ],
             ["visualizations-charts", {} ],
